server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
aiofiles==23.2.1
python-dotenv==1.0.1 